Edit Javafx files
