+++ build.gradle
@@ -10,7 +10,7 @@ repositories {
 }
 
 dependencies {
-    String javaFxVersion = '14'
+    String javaFxVersion = '11'
 
     implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'win'
     implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'mac'
@@ -43,11 +43,11 @@ test {
 }
 
 application {
-    mainClassName = "Duke.Duke"
+    mainClassName = "Launcher"
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "jarvis"
     archiveClassifier = null
 }
