--- src/main/java/Gui/Main.java
@@ -1,35 +1,0 @@
-package Gui;
-
-import java.io.IOException;
-
-import Duke.Duke;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
-/**
- * A GUI for Duke using FXML.
- */
-public class Main extends Application {
-
-    private Duke duke = new Duke();
-
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            // fxmlLoader.setController(this);
-            // fxmlLoader.setRoot(this);
-            AnchorPane ap = fxmlLoader.load();
-            // AnchorPane root = FXMLLoader.load(Main.class.getResource("/view/MainWindow.fxml"));
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
