+++ src/main/java/duke/Disk.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 import java.io.BufferedWriter;
 import java.io.File;
