+++ src/main/java/duke/Exceptions/EmptyMessageException.java
@@ -1,4 +1,4 @@
-package Duke.Exceptions;
+package duke.Exceptions;
 
 public class EmptyMessageException extends DukeException {
