+++ src/main/java/duke/Exceptions/MissingEnquiryDateException.java
@@ -1,4 +1,4 @@
-package Duke.Exceptions;
+package duke.Exceptions;
 
 public class MissingEnquiryDateException extends DukeException {
     /**
