+++ src/main/java/duke/ListStorage.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 import java.util.ArrayList;
 import java.util.List;
