+++ src/main/java/duke/Main.java
@@ -1,0 +1,24 @@
+package duke;
+
+import java.io.IOException;
+
+import duke.gui.MainWindow;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+
+/**
+ * A GUI for Duke using FXML.
+ */
+public class Main extends Application {
+
+    private Duke duke = new Duke();
+
+    @Override
+    public void start(Stage stage) {
+        MainWindow mainWindow = new MainWindow(stage);
+        mainWindow.setDuke(duke);
+    }
+}
