+++ src/main/java/duke/Parser.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 public class Parser {
     /**
