+++ src/main/java/duke/Task.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 import java.time.LocalDate;
