+++ src/main/java/duke/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package Gui;
+package duke.gui;
 import java.io.IOException;
 import java.util.Collections;
