+++ src/test/java/duke/ListStorageTest.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
