+++ src/test/java/duke/TaskTest.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
